Clean up Rust core: Clippy fixes and module extraction
- Use idiomatic slice params, is_some_and, is_none_or, or_default
- Implement FromStr trait for StorageHint instead of inherent from_str
- Replace manual range checks with contains()
- Add type aliases for boxed closures in view.rs (RowPredicate, ComputeFunction)
- Extract AggregateView support types into view/aggregate_support.rs
- Split python_bindings.rs into conversions.rs and iterators.rs submodules
- Update performance example to use renamed Table constructors

+++ impl/src/sequence.rs
@@ -441,7 +441,7 @@ mod tests {
 
         // Interleave appends and inserts to create non-uniform blocks
         for i in 0..50 {
-            seq.append((i * 2) as i32); // Even numbers
+            seq.append(i * 2); // Even numbers
         }
 
         // Insert odd numbers
